refactor: update import paths for hooks to use new naming convention
