+++ src/components/modules/report-issue/hooks/send-report-issue.hook.ts
@@ -6,7 +6,7 @@ import { z } from "zod";
 import { formSchema } from "../schema/report-issue.schema";
 import { addReportIssue } from "../server/report-issue.firebase";
 import { useGlobalAuthenticationStore } from "@/core/store/data";
-import { toast } from "@/hooks/use-toast";
+import { toast } from "@/hooks/toast.hook";
 import { IssueType } from "@/@types/issue.entity";
 
 export const useSendReportIssue = () => {
